migartion bug fix

+++ database/migrations/2020_12_17_135940_drop_contacts_table.php
@@ -13,7 +13,10 @@ class DropContactsTable extends Migration
      */
     public function up()
     {
-        Schema::drop('contacts');
+        if (Schema::hasTable('contacts')) {
+
+            Schema::drop('contacts');
+        }
     }
 
     /**
